extracting notifiable event fixtures to their own file
